requirements.txt
fastapi==0.115.12
uvicorn==0.34.2
APScheduler==3.10.4
sqlalchemy
python-dotenv
python-jose[cryptography]
passlib[bcrypt]
pydantic
pandas
openpyxl
psycopg2-binary
reportlab
pydantic-settings
httpx